Refactored main.js to libMain.js and commonFunctions.js
main.js was only used in library pages (not in standalone schedules or homepage), this caused issues where the programmer would add a function to main.js and then things got broken since it was never loaded and called nevertheless.

To fix this, the programmer created a new file called commonFunctions.js for storing functions that may be useful in all of the pages. He also renamed main.js to libMain.js in order to be less confused in the future.

The programmer also decided it would be better if all of the schedule day infos were capitalized. (Eg. Jämsä)